+++ kubernetes-client-openapi-discovery/src/main/resources/META-INF/services/io.micronaut.context.env.PropertySourceImporter
@@ -1,1 +1,1 @@
-io.micronaut.kubernetes.client.openapi.configuration.KubernetesPropertySourceImporter
+io.micronaut.kubernetes.client.openapi.configuration.imports.KubernetesPropertySourceImporter
